Flytt til egen pakke istedenfor felles
