+++ apps/aareg/build.gradle.kts
@@ -1,6 +1,7 @@
 val aaregClientVersion: String by project
 
 dependencies {
+    implementation(project(":kontrakt-domene-ansettelsesforhold"))
     implementation(project(":utils-auth"))
     implementation("no.nav.helsearbeidsgiver:aareg-client:$aaregClientVersion")
 }
