+++ apps/api/src/main/kotlin/no/nav/helsearbeidsgiver/inntektsmelding/api/response/Ansettelsesforhold.kt
@@ -4,7 +4,7 @@ package no.nav.helsearbeidsgiver.inntektsmelding.api.response
 
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.UseSerializers
-import no.nav.hag.simba.utils.felles.domene.Ansettelsesforhold
+import no.nav.hag.simba.kontrakt.domene.ansettelsesforhold.Ansettelsesforhold
 import no.nav.helsearbeidsgiver.utils.json.serializer.LocalDateSerializer
 import java.time.LocalDate
