+++ apps/api/src/test/kotlin/no/nav/helsearbeidsgiver/inntektsmelding/api/hentarbeidsforhold/HentArbeidsforholdRouteKtTest.kt
@@ -11,9 +11,9 @@ import io.mockk.coEvery
 import io.mockk.verify
 import io.mockk.verifySequence
 import kotlinx.serialization.json.JsonElement
+import no.nav.hag.simba.kontrakt.domene.ansettelsesforhold.Ansettelsesforhold
 import no.nav.hag.simba.utils.felles.EventName
 import no.nav.hag.simba.utils.felles.Key
-import no.nav.hag.simba.utils.felles.domene.Ansettelsesforhold
 import no.nav.hag.simba.utils.felles.domene.ResultJson
 import no.nav.hag.simba.utils.felles.json.toJson
 import no.nav.helsearbeidsgiver.inntektsmelding.api.RedisPoller
