+++ apps/faisu-service/src/test/kotlin/no/nav/helsearbeidsgiver/inntektsmelding/faisuservice/HentArbeidsforholdSelvbestemtServiceTest.kt
@@ -9,10 +9,10 @@ import io.mockk.mockk
 import io.mockk.verifySequence
 import kotlinx.serialization.builtins.MapSerializer
 import kotlinx.serialization.json.JsonElement
+import no.nav.hag.simba.kontrakt.domene.ansettelsesforhold.Ansettelsesforhold
 import no.nav.hag.simba.utils.felles.BehovType
 import no.nav.hag.simba.utils.felles.EventName
 import no.nav.hag.simba.utils.felles.Key
-import no.nav.hag.simba.utils.felles.domene.Ansettelsesforhold
 import no.nav.hag.simba.utils.felles.domene.ResultJson
 import no.nav.hag.simba.utils.felles.json.toJson
 import no.nav.hag.simba.utils.felles.test.json.lesBehov
