+++ utils/felles/src/main/kotlin/no/nav/hag/simba/utils/felles/utils/PeriodeUtils.kt
@@ -1,6 +1,5 @@
 package no.nav.hag.simba.utils.felles.utils
 
-import no.nav.hag.simba.utils.felles.domene.Ansettelsesforhold
 import no.nav.hag.simba.utils.felles.domene.PeriodeAapen
 import no.nav.helsearbeidsgiver.domene.inntektsmelding.v1.Periode
 import java.time.LocalDate
@@ -29,11 +28,4 @@ fun Periode.overlapperMed(annen: PeriodeAapen): Boolean {
     val endB = annen.tom ?: LocalDate.MAX
 
     return !this.fom.isAfter(endB) && !annen.fom.isAfter(endA)
-}
-
-fun Periode.overlapperMed(annen: Ansettelsesforhold): Boolean {
-    val endA = this.tom
-    val endB = annen.sluttdato ?: LocalDate.MAX
-
-    return !this.fom.isAfter(endB) && !annen.startdato.isAfter(endA)
 }
